+++ sample-app/src/components/CometChatHome/CometChatHome.tsx
@@ -1492,27 +1492,39 @@ function CometChatHome(props: { theme?: string }) {
 
         const openChatForUser = (user?: CometChat.User) => {
             const uid = user?.getUid();
-            if (uid) {
+            const closeSide = () => {
                 setAppState({ type: "updateSideComponent", payload: { visible: false, type: "" } });
+                setAppState({ type: 'updateThreadSearchMessage', payload: undefined });
+                setAppState({ type: 'updateThreadedMessage', payload: undefined });
+                setAppState({ type: 'updateGoToMessageId', payload: undefined });
+                setAppState({ type: "updateShowMessagesSearch", payload: false })
+            }
+            if (uid) {
                 if (activeTab === "chats") {
                     CometChat.getConversation(uid!, CometChatUIKitConstants.MessageReceiverType.user).then(
                         (conversation) => {
-                            setNewChat(undefined);
-                            setSelectedItem(conversation);
-                            setAppState({ type: "updateSelectedItem", payload: conversation });
+                            if(!selectedItem  || !(selectedItem instanceof CometChat.Conversation) || selectedItem?.getConversationId() !== conversation.getConversationId()) {
+                                setNewChat(undefined);
+                                setSelectedItem(conversation);
+                                setAppState({ type: "updateSelectedItem", payload: conversation });
+                                closeSide();
+                            }
                         },
                         (error) => {
                             setNewChat({ user, group: undefined });
                             setSelectedItem(undefined);
+                            closeSide();
                         }
                     );
                 } else if (activeTab === "users") {
                     setNewChat(undefined);
                     setSelectedItem(user);
                     setAppState({ type: "updateSelectedItemUser", payload: user });
+                    closeSide();
                 } else if (activeTab === "groups") {
                     setNewChat({ user, group: undefined });
                     setSelectedItem(undefined);
+                    closeSide();
                 }
             }
         }