+++ sample-app/src/metaInfo.ts
@@ -1,6 +1,6 @@
 export const metaInfo = {
   name: "sample-app",
-  version: "6.2.3",
+  version: "6.2.4",
   type: "sample",
   platform: "React",
 };